websocket/requirements.txt
fastapi
uvicorn[standard]
websockets
httpx
pandas
sqlalchemy
psycopg2-binary
asyncpg
msgpack
pydantic-settings
prometheus-client
python-dotenv
PyJWT[crypto]
cryptography
redis
pandas_ta
numpy
ostium-python-sdk
avantis-trader-sdk
lighter-sdk
langchain
langchain-openai
langchain-nvidia-ai-endpoints
langchain-community
deepagents>=0.3.5
fastembed
qdrant-client
langfuse
